chore: add delegate-stack-extend-increase (WIP)
